add LoginForSteamServerMessage

+++ SoulWorkerResearch.SoulCore/Extensions/BinaryReaderExtension.cs
@@ -9,44 +9,44 @@ public static class BinaryReaderExtension
 {
     #region Enumerable
 
-    public static IEnumerable<uint> ReadUInt32AsEnumerable(this BinaryReader reader, byte count) =>
+    public static IEnumerable<uint> ReadUInt32AsEnumerable(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadUInt32());
 
-    public static IEnumerable<int> ReadInt32AsEnumerable(this BinaryReader reader, byte count) =>
+    public static IEnumerable<int> ReadInt32AsEnumerable(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadInt32());
 
-    public static IEnumerable<ushort> ReadUInt16AsEnumerable(this BinaryReader reader, byte count) =>
+    public static IEnumerable<ushort> ReadUInt16AsEnumerable(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadUInt16());
 
-    public static IEnumerable<byte> ReadByteAsEnumerable(this BinaryReader reader, byte count) =>
+    public static IEnumerable<byte> ReadByteAsEnumerable(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadByte());
 
-    public static IEnumerable<BoosterEffectType> ReadBoosterEffectTypeAsEnumerable(this BinaryReader reader, byte count) =>
+    public static IEnumerable<BoosterEffectType> ReadBoosterEffectTypeAsEnumerable(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => (BoosterEffectType)reader.ReadByte());
 
-    public static IEnumerable<string> ReadByteLengthUnicodeStringAsEnumerable(this BinaryReader reader, byte count) =>
+    public static IEnumerable<string> ReadByteLengthUnicodeStringAsEnumerable(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadUTF16UnicodeString());
 
     #endregion Enumerable
 
     #region Arrays
 
-    public static uint[] ReadUInt32AsArray(this BinaryReader reader, byte count) =>
+    public static uint[] ReadUInt32AsArray(this BinaryReader reader, int count) =>
         reader.ReadUInt32AsEnumerable(count).ToArray();
 
-    public static int[] ReadInt32AsArray(this BinaryReader reader, byte count) =>
+    public static int[] ReadInt32AsArray(this BinaryReader reader, int count) =>
         reader.ReadInt32AsEnumerable(count).ToArray();
 
-    public static ushort[] ReadUInt16AsArray(this BinaryReader reader, byte count) =>
+    public static ushort[] ReadUInt16AsArray(this BinaryReader reader, int count) =>
         reader.ReadUInt16AsEnumerable(count).ToArray();
 
-    public static byte[] ReadByteAsArray(this BinaryReader reader, byte count) =>
+    public static byte[] ReadByteAsArray(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadByte()).ToArray();
 
-    public static float[] ReadSingleAsArray(this BinaryReader reader, byte count) =>
+    public static float[] ReadSingleAsArray(this BinaryReader reader, int count) =>
         Enumerable.Repeat(0, count).Select(_ => reader.ReadSingle()).ToArray();
 
-    public static string[] ReadByteLengthUnicodeStringAsArray(this BinaryReader reader, byte count) =>
+    public static string[] ReadByteLengthUnicodeStringAsArray(this BinaryReader reader, int count) =>
         reader.ReadByteLengthUnicodeStringAsEnumerable(count).ToArray();
 
     #endregion Arrays
